Actualizacion de Jonatan

+++ database/migrations/2016_01_15_214702_create_stocks_table.php
@@ -15,7 +15,7 @@ class CreateStocksTable extends Migration
         Schema::create('stocks', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('id_producto');
-            $table->decimal('cantidad');
+            $table->integer('cantidad');
             $table->timestamps();
         });
     }
